pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "os_heat_agent"
version = "0.1.0"
authors = [
  { name="Aurynn Shaw", email="aurynn@eiara.nz" },
]
description = "A small configurable agent for interacting with OpenStack Heat software configuration."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    # "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "boto3",
  "requests",
  "click",
  "structlog",
  "atpublic"
  # "schema"
]

[project.scripts]
os_heat_agent = "os_heat_agent:main"

[project.urls]
"Homepage" = "https://github.com/eiara/os-heat-agent"
"Bug Tracker" = "https://github.com/eiara/os-heat-agent/issues"

[tool.hatch.envs.default]
platforms = ["linux", "macos"]


[tool.hatch.envs.test]
platforms = ["linux", "macos"]
template = "default"
extra-dependencies = [
  "pytest"
]
[tool.hatch.envs.test.scripts]
# by default, disable integration tests
run = 'pytest -v -m "not integration"'
# Only run integration tests
integration = 'pytest -v -m "integration"'
all = 'pytest -v'

[tool.pytest.ini_options]
markers = [
  "integration: marks tests as integration tests, not run by default."
]
# So we can use the same-named files and functions in different directories, 
#   which otherwise doesn't work for historic reasons
# 
addopts = [
    "--import-mode=importlib",
]